# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/expenses": {
      "get": {
        "operationId": "ExpensesController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "expenses"
        ]
      },
      "post": {
        "operationId": "ExpensesController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "expenses"
        ]
      }
    },
    "/expenses/{id}": {
      "get": {
        "operationId": "ExpensesController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "expenses"
        ]
      },
      "patch": {
        "operationId": "ExpensesController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "expenses"
        ]
      },
      "delete": {
        "operationId": "ExpensesController_delete",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "expenses"
        ]
      }
    },
    "/categories": {
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_delete",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "categories"
        ]
      }
    },
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "notifications"
        ]
      },
      "post": {
        "operationId": "NotificationsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "get": {
        "operationId": "NotificationsController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "notifications"
        ]
      },
      "patch": {
        "operationId": "NotificationsController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "notifications"
        ]
      },
      "delete": {
        "operationId": "NotificationsController_delete",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/policies": {
      "get": {
        "operationId": "PoliciesController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "policies"
        ]
      },
      "post": {
        "operationId": "PoliciesController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "policies"
        ]
      }
    },
    "/policies/{id}": {
      "get": {
        "operationId": "PoliciesController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "policies"
        ]
      },
      "patch": {
        "operationId": "PoliciesController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "policies"
        ]
      },
      "delete": {
        "operationId": "PoliciesController_delete",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "policies"
        ]
      }
    },
...
  },
  "info": {
    "title": "FinStack API",
    "description": "REST API for FinStack expense, policy, audit, reporting, and dashboard workflows.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "RBAC role: superuser, admin, or user"
      }
    },
    "schemas": {
...
      "CreateExpenseDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "managerEmployeeId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "merchant": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "workflowStatus": {
            "type": "string",
            "enum": [
              "manager_review",
              "finance_review",
              "compliance_review",
              "approved_for_payment",
              "payment_processing",
              "returned",
              "rejected",
              "paid"
            ]
          },
          "notes": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "receiptFileName": {
            "type": "string"
          },
          "extraction_confidence": {
            "type": "number"
          },
          "flag": {
            "type": "string"
          },
          "risk_score": {
            "type": "number"
          }
        },
        "required": [
          "employeeId",
          "organizationId",
          "managerEmployeeId",
          "amount",
          "categoryId",
          "merchant",
          "date"
        ]
      },
      "UpdateExpenseDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "managerEmployeeId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "merchant": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "workflowStatus": {
            "type": "string",
            "enum": [
              "manager_review",
              "finance_review",
              "compliance_review",
              "approved_for_payment",
              "payment_processing",
              "returned",
              "rejected",
              "paid"
            ]
          },
          "notes": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "receiptFileName": {
            "type": "string"
          },
          "extraction_confidence": {
            "type": "number"
          },
          "flag": {
            "type": "string"
          },
          "risk_score": {
            "type": "number"
          },
          "managerDecision": {
            "type": "string"
          },
          "managerDecisionAt": {
            "type": "string"
          },
          "managerDecisionNote": {
            "type": "string"
          },
          "financeDecision": {
            "type": "string"
          },
          "financeDecisionAt": {
            "type": "string"
          },
          "financeDecisionNote": {
            "type": "string"
          },
          "complianceDecision": {
            "type": "string"
          },
          "complianceDecisionAt": {
            "type": "string"
          },
          "complianceDecisionNote": {
            "type": "string"
          },
          "escalatedByManager": {
            "type": "boolean"
          },
          "paidAt": {
            "type": "string"
          },
          "history": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "limit": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive"
            ]
          },
          "organizationId": {
            "type": "string"
          },
          "requiresReceipt": {
            "type": "boolean"
          },
          "color": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "limit",
          "organizationId"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "limit": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive"
            ]
          },
          "organizationId": {
            "type": "string"
          },
          "requiresReceipt": {
            "type": "boolean"
          },
          "color": {
            "type": "string"
          }
        }
      },
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "unread": {
            "type": "boolean"
          },
          "type": {
            "type": "string",
            "enum": [
              "info",
              "success",
              "warning",
              "danger"
            ]
          },
          "recipientEmployeeId": {
            "type": "string"
          },
          "recipientRole": {
            "type": "string",
            "enum": [
              "expense_submitter",
              "manager",
              "finance_officer",
              "compliance_officer",
              "configuration_manager",
              ""
            ]
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "relatedExpenseId": {
            "type": "string"
          },
          "relatedEntityId": {
            "type": "string"
          },
          "actionType": {
            "type": "string"
          },
          "dedupeKey": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "message"
        ]
      },
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {
          "unread": {
            "type": "boolean"
          },
          "type": {
            "type": "string",
            "enum": [
              "info",
              "success",
              "warning",
              "danger"
            ]
          },
          "recipientEmployeeId": {
            "type": "string"
          },
          "recipientRole": {
            "type": "string",
            "enum": [
              "expense_submitter",
              "manager",
              "finance_officer",
              "compliance_officer",
              "configuration_manager",
              ""
            ]
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "relatedExpenseId": {
            "type": "string"
          },
          "relatedEntityId": {
            "type": "string"
          },
          "actionType": {
            "type": "string"
          },
          "dedupeKey": {
            "type": "string"
          }
        }
      },
      "CreatePolicyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "maxAmount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "approval": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive"
            ]
          },
          "description": {
            "type": "string"
          },
          "requiresApproval": {
            "type": "boolean"
          },
          "receiptRequired": {
            "type": "boolean"
          },
          "organizationId": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "categoryId",
          "maxAmount",
          "organizationId"
        ]
      },
      "UpdatePolicyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "maxAmount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "approval": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive"
            ]
          },
          "description": {
            "type": "string"
          },
          "requiresApproval": {
            "type": "boolean"
          },
          "receiptRequired": {
            "type": "boolean"
          },
          "organizationId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}